Decoupling dependencies and add tests
